+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/direct/DirectBigQueryRelation.java
@@ -242,4 +242,9 @@ public class DirectBigQueryRelation extends BigQueryRelation
   public int hashCode() {
     return Objects.hash(getTableId(), schema(), compiledFilter);
   }
+
+  @Override
+  public String toString() {
+    return "DirectBigQueryRelation[" + toSqlTableReference(getTableId()) + "]";
+  }
 }